+++ app/src/main/java/cash/p/terminal/modules/send/bitcoin/advanced/SendBtcAdvancedSettingsScreen.kt
@@ -117,7 +117,7 @@ fun SendBtcAdvancedSettingsScreen(
                             errors = listOf(),
                             navController = fragmentNavController,
                             onValueChange = {
-                                sendBitcoinViewModel.updateFeeRate(it.toLong())
+                                sendBitcoinViewModel.updateFeeRate(it.toInt())
                             },
                             onClickIncrement = {
                                 sendBitcoinViewModel.incrementFeeRate()
